remove pubsub_*; remove worker; remove util::semantic_version;merge panic_hook into util; remove unused dependencies
