+++ .ci-scripts/loop_crates_to_run
@@ -97,8 +97,6 @@ function cargo_run_all () {
     cargo_run cita-common-benches
 
     cargo_run cita-types
-    cargo_run panic_hook
-    cargo_run worker
     cargo_run error
     cargo_run snappy
     cargo_run cita-directories
@@ -108,15 +106,7 @@ function cargo_run_all () {
 
     cargo_run authority_manage
 
-    cargo_run pubsub_rabbitmq
-    cargo_run pubsub_zeromq
-    cargo_run pubsub_kafka
-
-    for crate in pubsub; do
-        for features in ${FEATURES_FOR_PUBSUB}; do
-            cargo_run ${crate} --features "${features}"
-        done
-    done
+    cargo_run pubsub
 
     cargo_run blake2b
