+++ cita-ed25519/Cargo.toml
@@ -13,6 +13,8 @@ hashable = { path = "../hashable" }
 cita-crypto-trait = { path = "../cita-crypto-trait" }
 rlp = { path = "../rlp" }
 serde = "1.0"
+
+[dev-dependencies]
 bincode = "0.8.0"
 
 [features]
