+++ cita-secp256k1/Cargo.toml
@@ -15,6 +15,8 @@ cita-crypto-trait = { path = "../cita-crypto-trait" }
 rlp = { path = "../rlp" }
 rand = "0.3"
 serde = "1.0"
+
+[dev-dependencies]
 bincode = "0.8"
 
 [features]
