+++ libproto/Cargo.toml
@@ -18,7 +18,6 @@ rlp = { path = "../rlp" }
 serde = "1.0"
 serde_derive = "1.0"
 cita-logger = "0.1.0"
-tls-api = "0.1.14"
 
 [features]
 default = []
