--- panic_hook/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "panic_hook"
-version = "0.0.1"
-authors = ["Rivtower Technologies <contact@rivtower.com>"]
-license = "Apache-2.0"
-edition = "2018"
-
-[dependencies]
-backtrace = "0.3"
-cita-logger = "0.1.0"
