--- pubsub_kafka/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "pubsub_kafka"
-version = "0.1.0"
-authors = ["Rivtower Technologies <contact@rivtower.com>"]
-license = "Apache-2.0"
-edition = "2018"
-
-[dependencies.rdkafka]
-git = "https://github.com/fede1024/rust-rdkafka.git"
-rev = "84d4062"
-
-[dependencies]
-futures = "^0.1.13"
-cita-logger = "0.1.0"
-futures-cpupool = "0.1.6"
-crossbeam-channel = "0.3.8"
-
-[dev-dependencies]
-chrono = "^0.4.0"
-clap = "^2.18.0"
-futures-cpupool = "^0.1.2"
-rand = "^0.3.15"
-tokio-core = "^0.1.4"
