--- pubsub_rabbitmq/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "pubsub_rabbitmq"
-version = "0.1.0"
-authors = ["Rivtower Technologies <contact@rivtower.com>"]
-license = "Apache-2.0"
-edition = "2018"
-
-[dependencies.amqp]
-version = "0.1"
-default-features = false
-
-[dependencies]
-crossbeam-channel = "0.3.8"
