--- pubsub_zeromq/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "pubsub_zeromq"
-version = "0.1.0"
-authors = ["Rivtower Technologies <contact@rivtower.com>"]
-license = "Apache-2.0"
-edition = "2018"
-
-[dependencies.zmq]
-version = "0.8"
-default-features = false
-features = []
-
-[dependencies]
-cita-logger = "0.1.0"
-crossbeam-channel = "0.3.8"
