+++ util/Cargo.toml
@@ -6,18 +6,13 @@ license = "Apache-2.0"
 edition = "2018"
 
 [dependencies]
-rustc-hex = "1.0"
-elastic-array = "0.9"
-rlp = { path = "../rlp" }
-heapsize = "0.4"
 cita-types = { path = "../cita-types" }
-snappy = { path = "../snappy" }
 parking_lot = "0.6"
 ansi_term = "0.9"
 tiny-keccak= "1.4"
-lru-cache = "0.1.0"
-panic_hook = {path = "../panic_hook"}
 serde = "1.0"
 toml = "0.4"
 git2 = "0.7"
 rustc_version = "0.2.0"
+backtrace = "0.3"
+cita-logger = "0.1.0"
