--- worker/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "worker"
-version = "0.1.0"
-authors = ["Rivtower Technologies <contact@rivtower.com>"]
-license = "Apache-2.0"
-edition = "2018"
-
-[dependencies]
